pyproject.toml
[tool.poetry]
name = "one-code-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.3"
boto3 = "^1.16.25"
django-storages = "^1.10.1"
django-environ = "^0.4.5"
django-extensions = "^3.1.0"
django-cors-headers = "^3.5.0"
django-rest-auth = "^0.9.5"
djangorestframework = "^3.12.2"
python-slugify = "^4.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
